+++ packages/satusehat/src/Adaptor.js
@@ -200,18 +200,20 @@ export function patch(path, data, params = {}, callback = s => s) {
 }
 
 export {
-  fn,
-  fnIf,
   alterState,
   arrayToString,
+  as,
   combine,
   dataPath,
   dataValue,
   each,
   field,
   fields,
+  fn,
+  fnIf,
   http,
   lastReferenceValue,
+  map,
   merge,
   sourceValue,
   util,