+++ packages/whatsapp/src/Adaptor.js
@@ -1,0 +1,89 @@
+import { expandReferences } from '@openfn/language-common/util';
+import * as util from './Utils';
+
+/**
+ * State object
+ * @typedef {Object} HttpState
+ * @property data - the parsed response body
+ * @property response - the response from the HTTP server, including headers, statusCode, body, etc
+ * @property references - an array of all previous data objects used in the Job
+ **/
+
+/**
+ * Options provided to the HTTP request
+ * @typedef {Object} RequestOptions
+ * @public
+ * @property {object|string} body - body data to append to the request. JSON will be converted to a string (but a content-type header will not be attached to the request).
+ * @property {object} errors - Map of errorCodes -> error messages, ie, `{ 404: 'Resource not found;' }`. Pass `false` to suppress errors for this code.
+ * @property {object} form - Pass a JSON object to be serialised into a multipart HTML form (as FormData) in the body.
+ * @property {object} query - An object of query parameters to be encoded into the URL.
+ * @property {object} headers - An object of headers to append to the request.
+ * @property {string} parseAs - Parse the response body as json, text or stream. By default will use the response headers.
+ * @property {number} timeout - Request timeout in ms. Default: 300 seconds.
+ * @property {object} tls - TLS/SSL authentication options. See https://nodejs.org/api/tls.html#tlscreatesecurecontextoptions
+ */
+
+
+/**
+ * Make a HTTP request to the WhatsApp base URL, guided by the configuration
+ * @example
+ * <caption>Send a whatsapp message</caption>
+ * request(
+ *  'POST', 
+ *  'messages', 
+ *   { 
+ *    to: '254712345678', 
+ *    body: 'Hello, world!', 
+ *    type: 'template', 
+ *    template: { 
+ *      name: 'hello_world', 
+ *      language: { code: 'en_US' } 
+ *    }, 
+ *    messaging_product: 'whatsapp' 
+ * });
+ * @function
+ * @public
+ * @param {string} method - HTTP method to use
+ * @param {string} path - Path to resource
+ * @param {object} body - Object which will be attached to the request body
+ * @param {RequestOptions} options - Additional request options
+ * @returns {Operation}
+ * @state {HttpState}
+ */
+export function request(method, path, body, options = {}) {
+  return async state => {
+    const [resolvedMethod, resolvedPath, resolvedBody, resolvedoptions] =
+      expandReferences(state, method, path, body, options);
+
+    const response = await util.request(
+      state.configuration,
+      resolvedMethod,
+      resolvedPath,
+      {
+        body: resolvedBody,
+        ...resolvedoptions,
+      }
+    );
+
+    return util.prepareNextState(state, response);
+  };
+}
+
+export {
+  combine,
+  cursor,
+  dataPath,
+  dataValue,
+  dateFns,
+  each,
+  field,
+  fields,
+  fn,
+  fnIf,
+  group,
+  lastReferenceValue,
+  merge,
+  scrubEmojis,
+  sourceValue,
+  util,
+} from '@openfn/language-common';