+++ packages/fhir-4/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @openfn/language-fhir-4
 
+## 0.1.8 - 10 July 2025
+
+### Patch Changes
+
+- Updated dependencies \[cf9c09f]
+  - @openfn/language-common@3.0.1
+
+## 0.1.7 - 10 July 2025
+
+### Patch Changes
+
+- Updated dependencies \[ea85695]
+- Updated dependencies \[3fce58f]
+- Updated dependencies \[19f2d7e]
+- Updated dependencies \[f26bd2b]
+- Updated dependencies \[19f2d7e]
+  - @openfn/language-common@3.0.0
+
 ## 0.1.6 - 20 June 2025
 
 ### Patch Changes