+++ packages/salesforce/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @openfn/language-salesforce
 
+## 7.1.1 - 10 July 2025
+
+### Patch Changes
+
+- Updated dependencies \[cf9c09f]
+  - @openfn/language-common@3.0.1
+
+## 7.1.0 - 10 July 2025
+
+### Minor Changes
+
+- 19f2d7e: Exported `as()` function from common.
+- 8d78db4: Export `map()` function from common
+
+### Patch Changes
+
+- c4625fa: - Migrate from the deprecated `expandReferences` in `common` to the
+  new `expandReferences` from common.util
+  - Use `workspace:*` common version.
+- Updated dependencies \[ea85695]
+- Updated dependencies \[3fce58f]
+- Updated dependencies \[19f2d7e]
+- Updated dependencies \[f26bd2b]
+- Updated dependencies \[19f2d7e]
+  - @openfn/language-common@3.0.0
+
 ## 7.0.1 - 20 June 2025
 
 ### Patch Changes